@@ -5,11 +5,21 @@ set(CMAKE_PREFIX_PATH "${QTDIR}")
 set(CMAKE_INCLUDE_CURRENT_DIR ON)
 set(CMAKE_AUTOMOC ON)
 set(CMAKE_AUTOUIC ON)
-set (CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD 17)
 
 include(external/FindLibObs.cmake)
 
 find_package(LibObs REQUIRED)
+
+find_package(FFmpeg REQUIRED COMPONENTS avcodec avutil)
+# find_ffmpeg_library(avcodec)
+# find_ffmpeg_library(avutil)
+
+if(CMAKE_SIZEOF_VOID_P EQUAL 8)
+	set(ARCH 64)
+else()
+	set(ARCH 32)
+endif()
 
 if(WIN32)
 	find_library(WS2_32_LIBRARY ws2_32)
@@ -309,11 +386,14 @@ endif()
 
 # --- Linux-specific build settings and tasks ---
 if(UNIX AND NOT APPLE)
-	set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++11")
+	set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++17")
 
-	target_compile_options(mbedcrypto PRIVATE -fPIC)
-	set_target_properties(obs-ios-camera-source-win PROPERTIES PREFIX "")
-	target_link_libraries(obs-ios-camera-source-win
+	target_compile_options(portal PRIVATE -fPIC)
+	target_compile_options(libusbmuxd PRIVATE -fPIC)
+	target_compile_options(libplist PRIVATE -fPIC)
+	target_compile_options(libcnary PRIVATE -fPIC)
+	set_target_properties(obs-ios-camera-source PROPERTIES PREFIX "")
+	target_link_libraries(obs-ios-camera-source
 		obs-frontend-api)
 
 	file(GLOB locale_files data/locale/*.ini)
