@@ -18,7 +18,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-#include <stdlib.h>
+#include <cstdlib>
 #include <plist/Data.h>
 
 namespace PList
@@ -76,4 +76,4 @@ std::vector<char> Data::GetValue() const
 
 
 
-};
+}  // namespace PList
