@@ -18,7 +18,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-#include <stdlib.h>
+#include <cstdlib>
 #include <plist/Date.h>
 
 namespace PList
@@ -73,4 +73,4 @@ timeval Date::GetValue() const
     return t;
 }
 
-};
+}  // namespace PList
