@@ -18,7 +18,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-#include <stdlib.h>
+#include <cstdlib>
 #include <plist/Dictionary.h>
 
 namespace PList
@@ -28,28 +28,29 @@ Dictionary::Dictionary(Node* parent) : Structure(PLIST_DICT, parent)
 {
 }
 
-Dictionary::Dictionary(plist_t node, Node* parent) : Structure(parent)
+static void dictionary_fill(Dictionary *_this, std::map<std::string,Node*> &map, plist_t node)
 {
-    _node = node;
     plist_dict_iter it = NULL;
-
-    char* key = NULL;
+    plist_dict_new_iter(node, &it);
     plist_t subnode = NULL;
-    plist_dict_new_iter(_node, &it);
-    plist_dict_next_item(_node, it, &key, &subnode);
-    while (subnode)
-    {
-        _map[std::string(key)] = Node::FromPlist(subnode, this);
-
+    do {
+        char *key = NULL;
         subnode = NULL;
+        plist_dict_next_item(node, it, &key, &subnode);
+        if (key && subnode)
+            map[std::string(key)] = Node::FromPlist(subnode, _this);
         free(key);
-        key = NULL;
-        plist_dict_next_item(_node, it, &key, &subnode);
-    }
+    } while (subnode);
     free(it);
 }
 
-Dictionary::Dictionary(const PList::Dictionary& d) : Structure()
+Dictionary::Dictionary(plist_t node, Node* parent) : Structure(parent)
+{
+    _node = node;
+    dictionary_fill(this, _map, _node);
+}
+
+Dictionary::Dictionary(const PList::Dictionary& d)
 {
     for (Dictionary::iterator it = _map.begin(); it != _map.end(); it++)
     {
@@ -57,24 +58,8 @@ Dictionary::Dictionary(const PList::Dictionary& d) : Structure()
         delete it->second;
     }
     _map.clear();
-
     _node = plist_copy(d.GetPlist());
-    plist_dict_iter it = NULL;
-
-    char* key = NULL;
-    plist_t subnode = NULL;
-    plist_dict_new_iter(_node, &it);
-    plist_dict_next_item(_node, it, &key, &subnode);
-    while (subnode)
-    {
-        _map[std::string(key)] = Node::FromPlist(subnode, this);
-
-        subnode = NULL;
-        free(key);
-        key = NULL;
-        plist_dict_next_item(_node, it, &key, &subnode);
-    }
-    free(it);
+    dictionary_fill(this, _map, _node);
 }
 
 Dictionary& Dictionary::operator=(PList::Dictionary& d)
@@ -85,24 +70,8 @@ Dictionary& Dictionary::operator=(PList::Dictionary& d)
         delete it->second;
     }
     _map.clear();
-
     _node = plist_copy(d.GetPlist());
-    plist_dict_iter it = NULL;
-
-    char* key = NULL;
-    plist_t subnode = NULL;
-    plist_dict_new_iter(_node, &it);
-    plist_dict_next_item(_node, it, &key, &subnode);
-    while (subnode)
-    {
-        _map[std::string(key)] = Node::FromPlist(subnode, this);
-
-        subnode = NULL;
-        free(key);
-        key = NULL;
-        plist_dict_next_item(_node, it, NULL, &subnode);
-    }
-    free(it);
+    dictionary_fill(this, _map, _node);
     return *this;
 }
 
@@ -135,7 +104,22 @@ Dictionary::iterator Dictionary::End()
     return _map.end();
 }
 
+Dictionary::const_iterator Dictionary::Begin() const
+{
+    return _map.begin();
+}
+
+Dictionary::const_iterator Dictionary::End() const
+{
+    return _map.end();
+}
+
 Dictionary::iterator Dictionary::Find(const std::string& key)
+{
+    return _map.find(key);
+}
+
+Dictionary::const_iterator Dictionary::Find(const std::string& key) const
 {
     return _map.find(key);
 }
@@ -195,4 +179,4 @@ std::string Dictionary::GetNodeKey(Node* node)
     return "";
 }
 
-};
+}  // namespace PList
