@@ -18,7 +18,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-#include <stdlib.h>
+#include <cstdlib>
 #include <plist/String.h>
 
 namespace PList
@@ -77,4 +77,4 @@ std::string String::GetValue() const
     return ret;
 }
 
-};
+}  // namespace PList
