@@ -28,7 +28,7 @@ typedef struct bytearray_t {
 	size_t capacity;
 } bytearray_t;
 
-bytearray_t *byte_array_new();
+bytearray_t *byte_array_new(size_t initial);
 void byte_array_free(bytearray_t *ba);
 void byte_array_grow(bytearray_t *ba, size_t amount);
 void byte_array_append(bytearray_t *ba, void *buf, size_t len);
