@@ -25,10 +25,6 @@
 #include <config.h>
 #endif
 
-#ifdef _MSC_VER
-#include "msc_config.h"
-#endif
-
 #ifdef HAVE_STRPTIME
 #define _XOPEN_SOURCE 600
 #endif
@@ -40,11 +36,12 @@
 #include <time.h>
 
 #include <inttypes.h>
+#include <float.h>
 #include <math.h>
+#include <limits.h>
 
 #include <node.h>
 #include <node_list.h>
-#include <node_iterator.h>
 
 #include "plist.h"
 #include "base64.h"
@@ -73,6 +70,8 @@
 #define XPLIST_DICT_LEN 4
 
 #define MAC_EPOCH 978307200
+
+#define MAX_DATA_BYTES_PER_LINE(__i) (((76 - ((__i) << 3)) >> 2) * 3)
 
 static const char XML_PLIST_PROLOG[] = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n\
 <!DOCTYPE plist PUBLIC \"-//Apple//DTD PLIST 1.0//EN\" \"http://www.apple.com/DTDs/PropertyList-1.0.dtd\">\n\
@@ -104,33 +103,26 @@ void plist_xml_deinit(void)
 
 static size_t dtostr(char *buf, size_t bufsize, double realval)
 {
-    double f = realval;
-    double ip = 0.0;
-    int64_t v;
-    size_t len;
-    size_t p;
-
-    f = modf(f, &ip);
-    len = snprintf(buf, bufsize, "%s%" PRIi64, ((f < 0) && (ip >= 0)) ? "-" : "", (int64_t)ip);
-    if (len >= bufsize) {
-        return 0;
+    size_t len = 0;
+    if (isnan(realval)) {
+        len = snprintf(buf, bufsize, "nan");
+    } else if (isinf(realval)) {
+        len = snprintf(buf, bufsize, "%cinfinity", (realval > 0.0) ? '+' : '-');
+    } else if (realval == 0.0f) {
+        len = snprintf(buf, bufsize, "0.0");
+    } else {
+        size_t i = 0;
+        len = snprintf(buf, bufsize, "%.*g", 17, realval);
+        for (i = 0; i < len; i++) {
+            if (buf[i] == ',') {
+                buf[i] = '.';
+                break;
+            } else if (buf[i] == '.') {
+                break;
+            }
+        }
     }
-
-    if (f < 0) {
-        f *= -1;
-    }
-    f += 0.0000004;
-
-    p = len;
-    buf[p++] = '.';
-
-    while (p < bufsize && (p <= len+6)) {
-        f = modf(f*10, &ip);
-        v = (int)ip;
-        buf[p++] = (v + 0x30);
-    }
-    buf[p] = '\0';
-    return p;
+    return len;
 }
 
 static void node_to_xml(node_t* node, bytearray_t **outbuf, uint32_t depth)
@@ -204,12 +196,12 @@ static void node_to_xml(node_t* node, bytearray_t **outbuf, uint32_t depth)
     case PLIST_ARRAY:
         tag = XPLIST_ARRAY;
         tag_len = XPLIST_ARRAY_LEN;
-        isStruct = TRUE;
+        isStruct = (node->children) ? TRUE : FALSE;
         break;
     case PLIST_DICT:
         tag = XPLIST_DICT;
         tag_len = XPLIST_DICT_LEN;
-        isStruct = TRUE;
+        isStruct = (node->children) ? TRUE : FALSE;
         break;
     case PLIST_DATE:
         tag = XPLIST_DATE;
@@ -292,24 +284,24 @@ static void node_to_xml(node_t* node, bytearray_t **outbuf, uint32_t depth)
         tagOpen = TRUE;
         str_buf_append(*outbuf, "\n", 1);
         if (node_data->length > 0) {
-            char *buf = (char*)malloc(80);
             uint32_t j = 0;
             uint32_t indent = (depth > 8) ? 8 : depth;
-            uint32_t maxread = ((76 - indent*8) / 4) * 3;
+            uint32_t maxread = MAX_DATA_BYTES_PER_LINE(indent);
             size_t count = 0;
-            size_t b64count = 0;
-            str_buf_grow(*outbuf, (node_data->length / 3 * 4) + 4 + (((node_data->length / maxread) + 1) * (indent+1)));
+            size_t amount = (node_data->length / 3 * 4) + 4 + (((node_data->length / maxread) + 1) * (indent+1));
+            if ((*outbuf)->len + amount > (*outbuf)->capacity) {
+                str_buf_grow(*outbuf, amount);
+            }
             while (j < node_data->length) {
                 for (i = 0; i < indent; i++) {
                     str_buf_append(*outbuf, "\t", 1);
                 }
                 count = (node_data->length-j < maxread) ? node_data->length-j : maxread;
-                b64count = base64encode(buf, node_data->buff + j, count);
-                str_buf_append(*outbuf, buf, b64count);
+                assert((*outbuf)->len + count < (*outbuf)->capacity);
+                (*outbuf)->len += base64encode((char*)(*outbuf)->data + (*outbuf)->len, node_data->buff + j, count);
                 str_buf_append(*outbuf, "\n", 1);
                 j+=count;
             }
-            free(buf);
         }
         for (i = 0; i < depth; i++) {
             str_buf_append(*outbuf, "\t", 1);
@@ -352,21 +344,20 @@ static void node_to_xml(node_t* node, bytearray_t **outbuf, uint32_t depth)
     }
     free(val);
 
-    /* add return for structured types */
-    if (node_data->type == PLIST_ARRAY || node_data->type == PLIST_DICT)
+    if (isStruct) {
+        /* add newline for structured types */
         str_buf_append(*outbuf, "\n", 1);
 
-    if (isStruct) {
-        node_iterator_t *ni = node_iterator_create(node->children);
+        /* add child nodes */
+        if (node_data->type == PLIST_DICT && node->children) {
+            assert((node->children->count % 2) == 0);
+        }
         node_t *ch;
-        while ((ch = node_iterator_next(ni))) {
+        for (ch = node_first_child(node); ch; ch = node_next_sibling(ch)) {
             node_to_xml(ch, outbuf, depth+1);
         }
-        node_iterator_destroy(ni);
-    }
 
-    /* fix indent for structured types */
-    if (node_data->type == PLIST_ARRAY || node_data->type == PLIST_DICT) {
+        /* fix indent for structured types */
         for (i = 0; i < depth; i++) {
             str_buf_append(*outbuf, "\t", 1);
         }
@@ -379,8 +370,6 @@ static void node_to_xml(node_t* node, bytearray_t **outbuf, uint32_t depth)
         str_buf_append(*outbuf, ">", 1);
     }
     str_buf_append(*outbuf, "\n", 1);
-
-    return;
 }
 
 static void parse_date(const char *strval, struct TM *btime)
@@ -403,9 +392,131 @@ static void parse_date(const char *strval, struct TM *btime)
     btime->tm_isdst=0;
 }
 
+#define PO10i_LIMIT (INT64_MAX/10)
+
+/* based on https://stackoverflow.com/a/4143288 */
+static int num_digits_i(int64_t i)
+{
+    int n;
+    int64_t po10;
+    n=1;
+    if (i < 0) {
+        i = -i;
+        n++;
+    }
+    po10=10;
+    while (i>=po10) {
+        n++;
+        if (po10 > PO10i_LIMIT) break;
+        po10*=10;
+    }
+    return n;
+}
+
+#define PO10u_LIMIT (UINT64_MAX/10)
+
+/* based on https://stackoverflow.com/a/4143288 */
+static int num_digits_u(uint64_t i)
+{
+    int n;
+    uint64_t po10;
+    n=1;
+    po10=10;
+    while (i>=po10) {
+        n++;
+        if (po10 > PO10u_LIMIT) break;
+        po10*=10;
+    }
+    return n;
+}
+
+static void node_estimate_size(node_t *node, uint64_t *size, uint32_t depth)
+{
+    plist_data_t data;
+    if (!node) {
+        return;
+    }
+    data = plist_get_data(node);
+    if (node->children) {
+        node_t *ch;
+        for (ch = node_first_child(node); ch; ch = node_next_sibling(ch)) {
+            node_estimate_size(ch, size, depth + 1);
+        }
+        switch (data->type) {
+        case PLIST_DICT:
+            *size += (XPLIST_DICT_LEN << 1) + 7;
+            break;
+        case PLIST_ARRAY:
+            *size += (XPLIST_ARRAY_LEN << 1) + 7;
+            break;
+        default:
+            break;
+	}
+        *size += (depth << 1);
+    } else {
+        uint32_t indent = (depth > 8) ? 8 : depth;
+        switch (data->type) {
+        case PLIST_DATA: {
+            uint32_t req_lines = (data->length / MAX_DATA_BYTES_PER_LINE(indent)) + 1;
+            uint32_t b64len = data->length + (data->length / 3);
+            b64len += b64len % 4;
+            *size += b64len;
+            *size += (XPLIST_DATA_LEN << 1) + 5 + (indent+1) * (req_lines+1) + 1;
+        }   break;
+        case PLIST_STRING:
+            *size += data->length;
+            *size += (XPLIST_STRING_LEN << 1) + 6;
+            break;
+        case PLIST_KEY:
+            *size += data->length;
+            *size += (XPLIST_KEY_LEN << 1) + 6;
+            break;
+        case PLIST_UINT:
+            if (data->length == 16) {
+                *size += num_digits_u(data->intval);
+            } else {
+                *size += num_digits_i((int64_t)data->intval);
+            }
+            *size += (XPLIST_INT_LEN << 1) + 6;
+            break;
+        case PLIST_REAL:
+            *size += num_digits_i((int64_t)data->realval) + 7;
+            *size += (XPLIST_REAL_LEN << 1) + 6;
+            break;
+        case PLIST_DATE:
+            *size += 20; /* YYYY-MM-DDThh:mm:ssZ */
+            *size += (XPLIST_DATE_LEN << 1) + 6;
+            break;
+        case PLIST_BOOLEAN:
+            *size += ((data->boolval) ? XPLIST_TRUE_LEN : XPLIST_FALSE_LEN) + 4;
+            break;
+        case PLIST_DICT:
+            *size += XPLIST_DICT_LEN + 4; /* <dict/> */
+            break;
+        case PLIST_ARRAY:
+            *size += XPLIST_ARRAY_LEN + 4; /* <array/> */
+            break;
+        case PLIST_UID:
+            *size += num_digits_i((int64_t)data->intval);
+            *size += (XPLIST_DICT_LEN << 1) + 7;
+            *size += indent + ((indent+1) << 1);
+            *size += 18; /* <key>CF$UID</key> */
+            *size += (XPLIST_INT_LEN << 1) + 6;
+            break;
+        default:
+            break;
+        }
+        *size += indent;
+    }
+}
+
 PLIST_API void plist_to_xml(plist_t plist, char **plist_xml, uint32_t * length)
 {
-    strbuf_t *outbuf = str_buf_new();
+    uint64_t size = 0;
+    node_estimate_size((node_t*)plist, &size, 0);
+    size += sizeof(XML_PLIST_PROLOG) + sizeof(XML_PLIST_EPILOG) - 1;
+
+    strbuf_t *outbuf = str_buf_new(size);
 
     str_buf_append(outbuf, XML_PLIST_PROLOG, sizeof(XML_PLIST_PROLOG)-1);
 
@@ -418,6 +529,11 @@ PLIST_API void plist_to_xml(plist_t plist, char **plist_xml, uint32_t * length)
 
     outbuf->data = NULL;
     str_buf_free(outbuf);
+}
+
+PLIST_API void plist_to_xml_free(char *plist_xml)
+{
+    free(plist_xml);
 }
 
 struct _parse_ctx {
@@ -633,7 +749,7 @@ static text_part_t* get_text_parts(parse_ctx ctx, const char* tag, size_t tag_le
         }
     } while (1);
     ctx->pos++;
-    if (ctx->pos >= ctx->end-tag_len || strncmp(ctx->pos, tag, tag_len)) {
+    if (ctx->pos >= ctx->end-tag_len || strncmp(ctx->pos, tag, tag_len) != 0) {
         PLIST_XML_ERR("EOF or end tag mismatch\n");
         ctx->err++;
         return NULL;
@@ -849,7 +965,7 @@ static void node_from_xml(parse_ctx ctx, plist_t *plist)
                 ctx->err++;
                 goto err_out;
             }
-            if (strncmp(ctx->pos, "?>", 2)) {
+            if (strncmp(ctx->pos, "?>", 2) != 0) {
                 PLIST_XML_ERR("Couldn't find <? tag closing marker\n");
                 ctx->err++;
                 goto err_out;
@@ -861,7 +977,7 @@ static void node_from_xml(parse_ctx ctx, plist_t *plist)
             if (((ctx->end - ctx->pos) > 3) && !strncmp(ctx->pos, "!--", 3)) {
                 ctx->pos += 3;
                 find_str(ctx,"-->", 3, 0);
-                if (ctx->pos > ctx->end-3 || strncmp(ctx->pos, "-->", 3)) {
+                if (ctx->pos > ctx->end-3 || strncmp(ctx->pos, "-->", 3) != 0) {
                     PLIST_XML_ERR("Couldn't find end of comment\n");
                     ctx->err++;
                     goto err_out;
@@ -890,7 +1006,7 @@ static void node_from_xml(parse_ctx ctx, plist_t *plist)
                 }
                 if (embedded_dtd) {
                     find_str(ctx, "]>", 2, 1);
-                    if (ctx->pos > ctx->end-2 || strncmp(ctx->pos, "]>", 2)) {
+                    if (ctx->pos > ctx->end-2 || strncmp(ctx->pos, "]>", 2) != 0) {
                         PLIST_XML_ERR("Couldn't find end of DOCTYPE\n");
                         ctx->err++;
                         goto err_out;
@@ -1234,7 +1350,7 @@ static void node_from_xml(parse_ctx ctx, plist_t *plist)
                         break;
                     default:
                         /* should not happen */
-                        PLIST_XML_ERR("parent is not a structered node\n");
+                        PLIST_XML_ERR("parent is not a structured node\n");
                         ctx->err++;
                         goto err_out;
                     }
@@ -1304,11 +1420,6 @@ err_out:
         plist_free(*plist);
         *plist = NULL;
     }
-}
-
-PLIST_API void plist_to_xml_free(char **plist_xml)
-{
-	free(plist_xml);
 }
 
 PLIST_API void plist_from_xml(const char *plist_xml, uint32_t length, plist_t * plist)
