@@ -1,8 +1,8 @@
 /*
  * socket.c
  *
+ * Copyright (C) 2012-2020 Nikias Bassen <nikias@gmx.li>
  * Copyright (C) 2012 Martin Szulecki <m.szulecki@libimobiledevice.org>
- * Copyright (C) 2012 Nikias Bassen <nikias@gmx.li>
  *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
@@ -19,6 +19,9 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
+#ifdef HAVE_CONFIG_H
+#include <config.h>
+#endif
 #include <stdio.h>
 #include <stddef.h>
 #include <stdlib.h>
@@ -31,32 +34,98 @@
 #endif
 #ifdef WIN32
 #include <winsock2.h>
+#include <ws2tcpip.h>
 #include <windows.h>
 static int wsa_init = 0;
 #else
 #include <sys/socket.h>
 #include <sys/un.h>
 #include <netinet/in.h>
+#include <netinet/tcp.h>
 #include <netdb.h>
 #include <arpa/inet.h>
+#include <fcntl.h>
 #endif
 #include "socket.h"
 
 #define RECV_TIMEOUT 20000
+#define CONNECT_TIMEOUT 5000
+
+#ifndef EAFNOSUPPORT
+#define EAFNOSUPPORT 102
+#endif
+#ifndef ECONNRESET
+#define ECONNRESET 108
+#endif
+#ifndef ETIMEDOUT
+#define ETIMEDOUT 138
+#endif
 
 static int verbose = 0;
 
-void socket_set_verbose(int level)
+void usbmuxd_socket_set_verbose(int level)
 {
 	verbose = level;
 }
 
+const char *usbmuxd_socket_addr_to_string(struct sockaddr *addr, char *addr_out, size_t addr_out_size)
+{
+#ifdef WIN32
+	WSADATA wsa_data;
+	if (!wsa_init) {
+		if (WSAStartup(MAKEWORD(2,2), &wsa_data) != ERROR_SUCCESS) {
+			fprintf(stderr, "WSAStartup failed!\n");
+			ExitProcess(-1);
+		}
+		wsa_init = 1;
+	}
+	DWORD addr_out_len = addr_out_size;
+	DWORD addrlen = 0;
+
+	if (addr->sa_family == AF_INET) {
+		addrlen = sizeof(struct sockaddr_in);
+	}
+#ifdef AF_INET6
+	else if (addr->sa_family == AF_INET6) {
+		addrlen = sizeof(struct sockaddr_in6);
+	}
+#endif
+	else {
+		errno = EAFNOSUPPORT;
+		return NULL;
+	}
+
+	if (WSAAddressToString(addr, addrlen, NULL, addr_out, &addr_out_len) == 0) {
+		return addr_out;
+	}
+#else
+	const void *addrdata = NULL;
+
+	if (addr->sa_family == AF_INET) {
+		addrdata = &((struct sockaddr_in*)addr)->sin_addr;
+	}
+#ifdef AF_INET6
+	else if (addr->sa_family == AF_INET6) {
+		addrdata = &((struct sockaddr_in6*)addr)->sin6_addr;
+	}
+#endif
+	else {
+		errno = EAFNOSUPPORT;
+		return NULL;
+	}
+
+	if (inet_ntop(addr->sa_family, addrdata, addr_out, addr_out_size)) {
+		return addr_out;
+	}
+#endif
+	return NULL;
+}
+
 #ifndef WIN32
-int socket_create_unix(const char *filename)
+int usbmuxd_socket_create_unix(const char *filename)
 {
 	struct sockaddr_un name;
 	int sock;
-	size_t size;
 #ifdef SO_NOSIGPIPE
 	int yes = 1;
 #endif
@@ -65,7 +134,7 @@ int socket_create_unix(const char *filename)
 	unlink(filename);
 
 	/* Create the socket. */
-	sock = socket(PF_LOCAL, SOCK_STREAM, 0);
+	sock = socket(PF_UNIX, SOCK_STREAM, 0);
 	if (sock < 0) {
 		perror("socket");
 		return -1;
@@ -74,50 +143,40 @@ int socket_create_unix(const char *filename)
 #ifdef SO_NOSIGPIPE
 	if (setsockopt(sock, SOL_SOCKET, SO_NOSIGPIPE, (void*)&yes, sizeof(int)) == -1) {
 		perror("setsockopt()");
-		socket_close(sock);
+		usbmuxd_socket_close(sock);
 		return -1;
 	}
 #endif
 
 	/* Bind a name to the socket. */
-	name.sun_family = AF_LOCAL;
+	name.sun_family = AF_UNIX;
 	strncpy(name.sun_path, filename, sizeof(name.sun_path));
 	name.sun_path[sizeof(name.sun_path) - 1] = '\0';
 
-	/* The size of the address is
-	   the offset of the start of the filename,
-	   plus its length,
-	   plus one for the terminating null byte.
-	   Alternatively you can just do:
-	   size = SUN_LEN (&name);
-	 */
-	size = (offsetof(struct sockaddr_un, sun_path)
-			+ strlen(name.sun_path) + 1);
-
-	if (bind(sock, (struct sockaddr *) &name, size) < 0) {
+	if (bind(sock, (struct sockaddr*)&name, sizeof(name)) < 0) {
 		perror("bind");
-		socket_close(sock);
+		usbmuxd_socket_close(sock);
 		return -1;
 	}
 
 	if (listen(sock, 10) < 0) {
 		perror("listen");
-		socket_close(sock);
+		usbmuxd_socket_close(sock);
 		return -1;
 	}
 
 	return sock;
 }
 
-int socket_connect_unix(const char *filename)
+int usbmuxd_socket_connect_unix(const char *filename)
 {
 	struct sockaddr_un name;
 	int sfd = -1;
-	size_t size;
 	struct stat fst;
 #ifdef SO_NOSIGPIPE
 	int yes = 1;
 #endif
+	int bufsize = 0x20000;
 
 	// check if socket file exists...
 	if (stat(filename, &fst) != 0) {
@@ -134,33 +193,63 @@ int socket_connect_unix(const char *filename)
 		return -1;
 	}
 	// make a new socket
-	if ((sfd = socket(PF_LOCAL, SOCK_STREAM, 0)) < 0) {
+	if ((sfd = socket(PF_UNIX, SOCK_STREAM, 0)) < 0) {
 		if (verbose >= 2)
 			fprintf(stderr, "%s: socket: %s\n", __func__, strerror(errno));
 		return -1;
 	}
 
+	if (setsockopt(sfd, SOL_SOCKET, SO_SNDBUF, (void*)&bufsize, sizeof(int)) == -1) {
+		perror("Could not set send buffer for socket");
+	}
+
+	if (setsockopt(sfd, SOL_SOCKET, SO_RCVBUF, (void*)&bufsize, sizeof(int)) == -1) {
+		perror("Could not set receive buffer for socket");
+	}
+
 #ifdef SO_NOSIGPIPE
 	if (setsockopt(sfd, SOL_SOCKET, SO_NOSIGPIPE, (void*)&yes, sizeof(int)) == -1) {
 		perror("setsockopt()");
-		socket_close(sfd);
+		usbmuxd_socket_close(sfd);
 		return -1;
 	}
 #endif
-
 	// and connect to 'filename'
-	name.sun_family = AF_LOCAL;
+	name.sun_family = AF_UNIX;
 	strncpy(name.sun_path, filename, sizeof(name.sun_path));
 	name.sun_path[sizeof(name.sun_path) - 1] = 0;
 
-	size = (offsetof(struct sockaddr_un, sun_path)
-			+ strlen(name.sun_path) + 1);
+	int flags = fcntl(sfd, F_GETFL, 0);
+	fcntl(sfd, F_SETFL, flags | O_NONBLOCK);
 
-	if (connect(sfd, (struct sockaddr *) &name, size) < 0) {
-		socket_close(sfd);
+	do {
+		if (connect(sfd, (struct sockaddr*)&name, sizeof(name)) != -1) {
+			break;
+		}
+		if (errno == EINPROGRESS) {
+			fd_set fds;
+			FD_ZERO(&fds);
+			FD_SET(sfd, &fds);
+
+			struct timeval timeout;
+			timeout.tv_sec = CONNECT_TIMEOUT / 1000;
+			timeout.tv_usec = (CONNECT_TIMEOUT - (timeout.tv_sec * 1000)) * 1000;
+			if (select(sfd + 1, NULL, &fds, NULL, &timeout) == 1) {
+				int so_error;
+				socklen_t len = sizeof(so_error);
+				getsockopt(sfd, SOL_SOCKET, SO_ERROR, (void*)&so_error, &len);
+				if (so_error == 0) {
+					break;
+				}
+			}
+		}
+		usbmuxd_socket_close(sfd);
+		sfd = -1;
+	} while (0);
+
+	if (sfd < 0) {
 		if (verbose >= 2)
-			fprintf(stderr, "%s: connect: %s\n", __func__,
-					strerror(errno));
+			fprintf(stderr, "%s: connect: %s\n", __func__, strerror(errno));
 		return -1;
 	}
 
@@ -168,7 +257,7 @@ int socket_connect_unix(const char *filename)
 }
 #endif
 
-int socket_create(uint16_t port)
+int usbmuxd_socket_create(uint16_t port)
 {
 	int sfd = -1;
 	int yes = 1;
@@ -191,14 +280,14 @@ int socket_create(uint16_t port)
 
 	if (setsockopt(sfd, SOL_SOCKET, SO_REUSEADDR, (void*)&yes, sizeof(int)) == -1) {
 		perror("setsockopt()");
-		socket_close(sfd);
+		usbmuxd_socket_close(sfd);
 		return -1;
 	}
 
 #ifdef SO_NOSIGPIPE
 	if (setsockopt(sfd, SOL_SOCKET, SO_NOSIGPIPE, (void*)&yes, sizeof(int)) == -1) {
 		perror("setsockopt()");
-		socket_close(sfd);
+		usbmuxd_socket_close(sfd);
 		return -1;
 	}
 #endif
@@ -210,26 +299,27 @@ int socket_create(uint16_t port)
 
 	if (0 > bind(sfd, (struct sockaddr *) &saddr, sizeof(saddr))) {
 		perror("bind()");
-		socket_close(sfd);
+		usbmuxd_socket_close(sfd);
 		return -1;
 	}
 
 	if (listen(sfd, 1) == -1) {
 		perror("listen()");
-		socket_close(sfd);
+		usbmuxd_socket_close(sfd);
 		return -1;
 	}
 
 	return sfd;
 }
 
-int socket_connect(const char *addr, uint16_t port)
+int usbmuxd_socket_connect_addr(struct sockaddr* addr, uint16_t port)
 {
 	int sfd = -1;
 	int yes = 1;
-	struct hostent *hp;
-	struct sockaddr_in saddr;
+	int bufsize = 0x20000;
+	int addrlen = 0;
 #ifdef WIN32
+	u_long l_yes = 1;
 	WSADATA wsa_data;
 	if (!wsa_init) {
 		if (WSAStartup(MAKEWORD(2,2), &wsa_data) != ERROR_SUCCESS) {
@@ -240,58 +330,227 @@ int socket_connect(const char *addr, uint16_t port)
 	}
 #endif
 
-	if (!addr) {
-		errno = EINVAL;
+	if (addr->sa_family == AF_INET) {
+		struct sockaddr_in* addr_in = (struct sockaddr_in*)addr;
+		addr_in->sin_port = htons(port);
+		addrlen = sizeof(struct sockaddr_in);
+	}
+#ifdef AF_INET6
+	else if (addr->sa_family == AF_INET6) {
+		struct sockaddr_in6* addr_in = (struct sockaddr_in6*)addr;
+		addr_in->sin6_port = htons(port);
+		addrlen = sizeof(struct sockaddr_in6);
+	}
+#endif
+	else {
+		fprintf(stderr, "ERROR: Unsupported address family");
 		return -1;
 	}
 
-	if ((hp = gethostbyname(addr)) == NULL) {
-		if (verbose >= 2)
-			fprintf(stderr, "%s: unknown host '%s'\n", __func__, addr);
-		return -1;
-	}
-
-	if (!hp->h_addr) {
-		if (verbose >= 2)
-			fprintf(stderr, "%s: gethostbyname returned NULL address!\n",
-					__func__);
-		return -1;
-	}
-
-	if (0 > (sfd = socket(PF_INET, SOCK_STREAM, IPPROTO_TCP))) {
+	sfd = socket(addr->sa_family, SOCK_STREAM, IPPROTO_TCP);
+	if (sfd == -1) {
 		perror("socket()");
-		return -1;
-	}
-
-	if (setsockopt(sfd, SOL_SOCKET, SO_REUSEADDR, (void*)&yes, sizeof(int)) == -1) {
-		perror("setsockopt()");
-		socket_close(sfd);
 		return -1;
 	}
 
 #ifdef SO_NOSIGPIPE
 	if (setsockopt(sfd, SOL_SOCKET, SO_NOSIGPIPE, (void*)&yes, sizeof(int)) == -1) {
 		perror("setsockopt()");
-		socket_close(sfd);
+		usbmuxd_socket_close(sfd);
 		return -1;
 	}
 #endif
 
-	memset((void *) &saddr, 0, sizeof(saddr));
-	saddr.sin_family = AF_INET;
-	saddr.sin_addr.s_addr = *(uint32_t *) hp->h_addr;
-	saddr.sin_port = htons(port);
+	if (setsockopt(sfd, SOL_SOCKET, SO_REUSEADDR, (void*)&yes, sizeof(int)) == -1) {
+		perror("setsockopt()");
+		usbmuxd_socket_close(sfd);
+		return -1;
+	}
 
-	if (connect(sfd, (struct sockaddr *) &saddr, sizeof(saddr)) < 0) {
-		perror("connect");
-		socket_close(sfd);
-		return -2;
+#ifdef WIN32
+	ioctlsocket(sfd, FIONBIO, &l_yes);
+#else
+	int flags = fcntl(sfd, F_GETFL, 0);
+	fcntl(sfd, F_SETFL, flags | O_NONBLOCK);
+#endif
+
+	do {
+		if (connect(sfd, addr, addrlen) != -1) {
+			break;
+		}
+#ifdef WIN32
+		if (WSAGetLastError() == WSAEWOULDBLOCK)
+#else
+		if (errno == EINPROGRESS)
+#endif
+		{
+			fd_set fds;
+			FD_ZERO(&fds);
+			FD_SET(sfd, &fds);
+
+			struct timeval timeout;
+			timeout.tv_sec = CONNECT_TIMEOUT / 1000;
+			timeout.tv_usec = (CONNECT_TIMEOUT - (timeout.tv_sec * 1000)) * 1000;
+			if (select(sfd + 1, NULL, &fds, NULL, &timeout) == 1) {
+				int so_error;
+				socklen_t len = sizeof(so_error);
+				getsockopt(sfd, SOL_SOCKET, SO_ERROR, (void*)&so_error, &len);
+				if (so_error == 0) {
+					errno = 0;
+					break;
+				}
+				errno = so_error;
+			}
+		}
+		usbmuxd_socket_close(sfd);
+		sfd = -1;
+	} while (0);
+
+	if (sfd < 0) {
+		if (verbose >= 2) {
+			char addrtxt[48];
+			usbmuxd_socket_addr_to_string(addr, addrtxt, sizeof(addrtxt));
+			fprintf(stderr, "%s: Could not connect to %s port %d\n", __func__, addrtxt, port);
+		}
+		return -1;
+	}
+
+	if (setsockopt(sfd, IPPROTO_TCP, TCP_NODELAY, (void*)&yes, sizeof(int)) == -1) {
+		perror("Could not set TCP_NODELAY on socket");
+	}
+
+	if (setsockopt(sfd, SOL_SOCKET, SO_SNDBUF, (void*)&bufsize, sizeof(int)) == -1) {
+		perror("Could not set send buffer for socket");
+	}
+
+	if (setsockopt(sfd, SOL_SOCKET, SO_RCVBUF, (void*)&bufsize, sizeof(int)) == -1) {
+		perror("Could not set receive buffer for socket");
 	}
 
 	return sfd;
 }
 
-int socket_check_fd(int fd, fd_mode fdm, unsigned int timeout)
+int usbmuxd_socket_connect(const char *addr, uint16_t port)
+{
+	int sfd = -1;
+	int yes = 1;
+	int bufsize = 0x20000;
+	struct addrinfo hints;
+	struct addrinfo *result, *rp;
+	char portstr[8];
+	int res;
+#ifdef WIN32
+	u_long l_yes = 1;
+	WSADATA wsa_data;
+	if (!wsa_init) {
+		if (WSAStartup(MAKEWORD(2,2), &wsa_data) != ERROR_SUCCESS) {
+			fprintf(stderr, "WSAStartup failed!\n");
+			ExitProcess(-1);
+		}
+		wsa_init = 1;
+	}
+#else
+	int flags = 0;
+#endif
+
+	if (!addr) {
+		errno = EINVAL;
+		return -1;
+	}
+
+	memset(&hints, '\0', sizeof(struct addrinfo));
+	hints.ai_family = AF_UNSPEC;
+	hints.ai_socktype = SOCK_STREAM;
+	hints.ai_flags = 0;
+	hints.ai_protocol = IPPROTO_TCP;
+
+	sprintf(portstr, "%d", port);
+
+	res = getaddrinfo(addr, portstr, &hints, &result);
+	if (res != 0) {
+		fprintf(stderr, "%s: getaddrinfo: %s\n", __func__, gai_strerror(res));
+		return -1;
+	}
+
+	for (rp = result; rp != NULL; rp = rp->ai_next) {
+		sfd = socket(rp->ai_family, rp->ai_socktype, rp->ai_protocol);
+		if (sfd == -1) {
+			continue;
+		}
+
+#ifdef SO_NOSIGPIPE
+		if (setsockopt(sfd, SOL_SOCKET, SO_NOSIGPIPE, (void*)&yes, sizeof(int)) == -1) {
+			perror("setsockopt()");
+			usbmuxd_socket_close(sfd);
+			return -1;
+		}
+#endif
+
+		if (setsockopt(sfd, SOL_SOCKET, SO_REUSEADDR, (void*)&yes, sizeof(int)) == -1) {
+			perror("setsockopt()");
+			usbmuxd_socket_close(sfd);
+			continue;
+		}
+
+#ifdef WIN32
+		ioctlsocket(sfd, FIONBIO, &l_yes);
+#else
+		flags = fcntl(sfd, F_GETFL, 0);
+		fcntl(sfd, F_SETFL, flags | O_NONBLOCK);
+#endif
+
+		if (connect(sfd, rp->ai_addr, rp->ai_addrlen) != -1) {
+			break;
+		}
+#ifdef WIN32
+		if (WSAGetLastError() == WSAEWOULDBLOCK)
+#else
+		if (errno == EINPROGRESS)
+#endif
+		{
+			fd_set fds;
+			FD_ZERO(&fds);
+			FD_SET(sfd, &fds);
+
+			struct timeval timeout;
+			timeout.tv_sec = CONNECT_TIMEOUT / 1000;
+			timeout.tv_usec = (CONNECT_TIMEOUT - (timeout.tv_sec * 1000)) * 1000;
+			if (select(sfd + 1, NULL, &fds, NULL, &timeout) == 1) {
+				int so_error;
+				socklen_t len = sizeof(so_error);
+				getsockopt(sfd, SOL_SOCKET, SO_ERROR, (void*)&so_error, &len);
+				if (so_error == 0) {
+					break;
+				}
+			}
+		}
+		usbmuxd_socket_close(sfd);
+	}
+
+	freeaddrinfo(result);
+
+	if (rp == NULL) {
+		if (verbose >= 2)
+			fprintf(stderr, "%s: Could not connect to %s:%d\n", __func__, addr, port);
+		return -1;
+	}
+
+	if (setsockopt(sfd, IPPROTO_TCP, TCP_NODELAY, (void*)&yes, sizeof(int)) == -1) {
+		perror("Could not set TCP_NODELAY on socket");
+	}
+
+	if (setsockopt(sfd, SOL_SOCKET, SO_SNDBUF, (void*)&bufsize, sizeof(int)) == -1) {
+		perror("Could not set send buffer for socket");
+	}
+
+	if (setsockopt(sfd, SOL_SOCKET, SO_RCVBUF, (void*)&bufsize, sizeof(int)) == -1) {
+		perror("Could not set receive buffer for socket");
+	}
+
+	return sfd;
+}
+
+int usbmuxd_socket_check_fd(int fd, fd_mode fdm, unsigned int timeout)
 {
 	fd_set fds;
 	int sret;
@@ -308,17 +567,16 @@ int socket_check_fd(int fd, fd_mode fdm, unsigned int timeout)
 	FD_ZERO(&fds);
 	FD_SET(fd, &fds);
 
-	if (timeout > 0) {
-		to.tv_sec = (time_t) (timeout / 1000);
-		to.tv_usec = (time_t) ((timeout - (to.tv_sec * 1000)) * 1000);
-		pto = &to;
-	} else {
-		pto = NULL;
-	}
-
 	sret = -1;
 
 	do {
+		if (timeout > 0) {
+			to.tv_sec = (time_t) (timeout / 1000);
+			to.tv_usec = (time_t) ((timeout - (to.tv_sec * 1000)) * 1000);
+			pto = &to;
+		} else {
+			pto = NULL;
+		}
 		eagain = 0;
 		switch (fdm) {
 		case FDM_READ:
@@ -352,13 +610,17 @@ int socket_check_fd(int fd, fd_mode fdm, unsigned int timeout)
 							strerror(errno));
 				return -1;
 			}
+		} else if (sret == 0) {
+			if (verbose >= 2)
+				fprintf(stderr, "%s: timeout\n", __func__);
+			return -ETIMEDOUT;
 		}
 	} while (eagain);
 
 	return sret;
 }
 
-int socket_accept(int fd, uint16_t port)
+int usbmuxd_socket_accept(int fd, uint16_t port)
 {
 #ifdef WIN32
 	int addr_len;
@@ -379,12 +641,12 @@ int socket_accept(int fd, uint16_t port)
 	return result;
 }
 
-int socket_shutdown(int fd, int how)
+int usbmuxd_socket_shutdown(int fd, int how)
 {
 	return shutdown(fd, how);
 }
 
-int socket_close(int fd) {
+int usbmuxd_socket_close(int fd) {
 #ifdef WIN32
 	return closesocket(fd);
 #else
@@ -392,24 +654,24 @@ int socket_close(int fd) {
 #endif
 }
 
-int socket_receive(int fd, void *data, size_t length)
+int usbmuxd_socket_receive(int fd, void *data, size_t length)
 {
-	return socket_receive_timeout(fd, data, length, 0, RECV_TIMEOUT);
+	return usbmuxd_socket_receive_timeout(fd, data, length, 0, RECV_TIMEOUT);
 }
 
-int socket_peek(int fd, void *data, size_t length)
+int usbmuxd_socket_peek(int fd, void *data, size_t length)
 {
-	return socket_receive_timeout(fd, data, length, MSG_PEEK, RECV_TIMEOUT);
+	return usbmuxd_socket_receive_timeout(fd, data, length, MSG_PEEK, RECV_TIMEOUT);
 }
 
-int socket_receive_timeout(int fd, void *data, size_t length, int flags,
+int usbmuxd_socket_receive_timeout(int fd, void *data, size_t length, int flags,
 					 unsigned int timeout)
 {
 	int res;
 	int result;
 
 	// check if data is available
-	res = socket_check_fd(fd, FDM_READ, timeout);
+	res = usbmuxd_socket_check_fd(fd, FDM_READ, timeout);
 	if (res <= 0) {
 		return res;
 	}
@@ -419,7 +681,7 @@ int socket_receive_timeout(int fd, void *data, size_t length, int flags,
 		// but this is an error condition
 		if (verbose >= 3)
 			fprintf(stderr, "%s: fd=%d recv returned 0\n", __func__, fd);
-		return -EAGAIN;
+		return -ECONNRESET;
 	}
 	if (result < 0) {
 #ifdef WIN32
@@ -431,9 +693,13 @@ int socket_receive_timeout(int fd, void *data, size_t length, int flags,
 	return result;
 }
 
-int socket_send(int fd, void *data, size_t length)
+int usbmuxd_socket_send(int fd, void *data, size_t length)
 {
 	int flags = 0;
+	int res = usbmuxd_socket_check_fd(fd, FDM_WRITE, 1000);
+	if (res <= 0) {
+		return res;
+	}
 #ifdef MSG_NOSIGNAL
 	flags |= MSG_NOSIGNAL;
 #endif
