@@ -1,8 +1,8 @@
 /*
  * socket.h
  *
+ * Copyright (C) 2012-2020 Nikias Bassen <nikias@gmx.li>
  * Copyright (C) 2012 Martin Szulecki <m.szulecki@libimobiledevice.org>
- * Copyright (C) 2012 Nikias Bassen <nikias@gmx.li>
  *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
@@ -42,24 +42,27 @@ typedef enum fd_mode fd_mode;
 #endif
 
 #ifndef WIN32
-int socket_create_unix(const char *filename);
-int socket_connect_unix(const char *filename);
+int usbmuxd_socket_create_unix(const char *filename);
+int usbmuxd_socket_connect_unix(const char *filename);
 #endif
-int socket_create(uint16_t port);
-int socket_connect(const char *addr, uint16_t port);
-int socket_check_fd(int fd, fd_mode fdm, unsigned int timeout);
-int socket_accept(int fd, uint16_t port);
+int usbmuxd_socket_create(uint16_t port);
+int usbmuxd_socket_connect_addr(struct sockaddr *addr, uint16_t port);
+int usbmuxd_socket_connect(const char *addr, uint16_t port);
+int usbmuxd_socket_check_fd(int fd, fd_mode fdm, unsigned int timeout);
+int usbmuxd_socket_accept(int fd, uint16_t port);
 
-int socket_shutdown(int fd, int how);
-int socket_close(int fd);
+int usbmuxd_socket_shutdown(int fd, int how);
+int usbmuxd_socket_close(int fd);
 
-int socket_receive(int fd, void *data, size_t size);
-int socket_peek(int fd, void *data, size_t size);
-int socket_receive_timeout(int fd, void *data, size_t size, int flags,
+int usbmuxd_socket_receive(int fd, void *data, size_t size);
+int usbmuxd_socket_peek(int fd, void *data, size_t size);
+int usbmuxd_socket_receive_timeout(int fd, void *data, size_t size, int flags,
 					 unsigned int timeout);
 
-int socket_send(int fd, void *data, size_t size);
+int usbmuxd_socket_send(int fd, void *data, size_t size);
 
-void socket_set_verbose(int level);
+void usbmuxd_socket_set_verbose(int level);
+
+const char *usbmuxd_socket_addr_to_string(struct sockaddr *addr, char *addr_out, size_t addr_out_size);
 
 #endif	/* SOCKET_SOCKET_H */
