@@ -1,9 +1,9 @@
 /*
  * iproxy.c -- proxy that enables tcp service access to iOS devices
  *
- * Copyright (C) 2014	Martin Szulecki <m.szulecki@libimobiledevice.org>
- * Copyright (C) 2009	Nikias Bassen <nikias@gmx.li>
- * Copyright (C) 2009	Paul Sladen <libiphone@paul.sladen.org>
+ * Copyright (C) 2009-2020 Nikias Bassen <nikias@gmx.li>
+ * Copyright (C) 2014      Martin Szulecki <m.szulecki@libimobiledevice.org>
+ * Copyright (C) 2009      Paul Sladen <libiphone@paul.sladen.org>
  *
  * Based upon iTunnel source code, Copyright (c) 2008 Jing Su.
  * http://www.cs.toronto.edu/~jingsu/itunnel/
@@ -23,12 +23,16 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
+#ifdef HAVE_CONFIG_H
+#include <config.h>
+#endif
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 #include <fcntl.h>
 #include <stddef.h>
 #include <errno.h>
+#include <getopt.h>
 #ifdef WIN32
 #include <winsock2.h>
 #include <windows.h>
@@ -39,9 +43,16 @@ typedef unsigned int socklen_t;
 #include <arpa/inet.h>
 #include <pthread.h>
 #include <netinet/in.h>
+#include <signal.h>
 #endif
 #include "socket.h"
 #include "usbmuxd.h"
+
+#ifndef ETIMEDOUT
+#define ETIMEDOUT 138
+#endif
+
+static int debug_level = 0;
 
 #ifdef _MSC_VER
 #define strcasecmp stricmp
@@ -50,6 +61,7 @@ typedef unsigned int socklen_t;
 static uint16_t listen_port = 0;
 static uint16_t device_port = 0;
 static char* device_udid = NULL;
+static enum usbmux_lookup_options lookup_opts = 0;
 
 struct client_data {
 	int fd;
@@ -70,9 +82,9 @@ static void *run_stoc_loop(void *arg)
 	printf("%s: fd = %d\n", __func__, cdata->fd);
 
 	while (!cdata->stop_stoc && cdata->fd > 0 && cdata->sfd > 0) {
-		recv_len = socket_receive_timeout(cdata->sfd, buffer, sizeof(buffer), 0, 5000);
+		recv_len = usbmuxd_socket_receive_timeout(cdata->sfd, buffer, sizeof(buffer), 0, 5000);
 		if (recv_len <= 0) {
-			if (recv_len == 0) {
+			if (recv_len == 0 || recv_len == -ETIMEDOUT) {
 				// try again
 				continue;
 			} else {
@@ -81,7 +93,7 @@ static void *run_stoc_loop(void *arg)
 			}
 		} else {
 			// send to socket
-			sent = socket_send(cdata->fd, buffer, recv_len);
+			sent = usbmuxd_socket_send(cdata->fd, buffer, recv_len);
 			if (sent < recv_len) {
 				if (sent <= 0) {
 					fprintf(stderr, "send failed: %s\n", strerror(errno));
@@ -93,7 +105,7 @@ static void *run_stoc_loop(void *arg)
 		}
 	}
 
-	socket_close(cdata->fd);
+	usbmuxd_socket_close(cdata->fd);
 
 	cdata->fd = -1;
 	cdata->stop_ctos = 1;
@@ -123,9 +135,9 @@ static void *run_ctos_loop(void *arg)
 #endif
 
 	while (!cdata->stop_ctos && cdata->fd>0 && cdata->sfd>0) {
-		recv_len = socket_receive_timeout(cdata->fd, buffer, sizeof(buffer), 0, 5000);
+		recv_len = usbmuxd_socket_receive_timeout(cdata->fd, buffer, sizeof(buffer), 0, 5000);
 		if (recv_len <= 0) {
-			if (recv_len == 0) {
+			if (recv_len == 0 || recv_len == -ETIMEDOUT) {
 				// try again
 				continue;
 			} else {
@@ -134,7 +146,7 @@ static void *run_ctos_loop(void *arg)
 			}
 		} else {
 			// send to local socket
-			sent = socket_send(cdata->sfd, buffer, recv_len);
+			sent = usbmuxd_socket_send(cdata->sfd, buffer, recv_len);
 			if (sent < recv_len) {
 				if (sent <= 0) {
 					fprintf(stderr, "send failed: %s\n", strerror(errno));
@@ -146,7 +158,7 @@ static void *run_ctos_loop(void *arg)
 		}
 	}
 
-	socket_close(cdata->fd);
+	usbmuxd_socket_close(cdata->fd);
 
 	cdata->fd = -1;
 	cdata->stop_stoc = 1;
@@ -164,6 +176,8 @@ static void *acceptor_thread(void *arg)
 {
 	struct client_data *cdata;
 	usbmuxd_device_info_t *dev_list = NULL;
+	usbmuxd_device_info_t *dev = NULL;
+	usbmuxd_device_info_t muxdev;
 #ifdef WIN32
 	HANDLE ctos = NULL;
 #else
@@ -178,57 +192,92 @@ static void *acceptor_thread(void *arg)
 
 	cdata = (struct client_data*)arg;
 
-	if ((count = usbmuxd_get_device_list(&dev_list)) < 0) {
-		printf("Connecting to usbmuxd failed, terminating.\n");
-		free(dev_list);
-		if (cdata->fd > 0) {
-			socket_close(cdata->fd);
-		}
-		free(cdata);
-		return NULL;
-	}
-
-	fprintf(stdout, "Number of available devices == %d\n", count);
-
-	if (dev_list == NULL || dev_list[0].handle == 0) {
-		printf("No connected device found, terminating.\n");
-		free(dev_list);
-		if (cdata->fd > 0) {
-			socket_close(cdata->fd);
-		}
-		free(cdata);
-		return NULL;
-	}
-
-	usbmuxd_device_info_t *dev = NULL;
 	if (device_udid) {
+		if (usbmuxd_get_device(device_udid, &muxdev, lookup_opts) > 0) {
+			dev = &muxdev;
+		}
+	} else {
+		if ((count = usbmuxd_get_device_list(&dev_list)) < 0) {
+			printf("Connecting to usbmuxd failed, terminating.\n");
+			free(dev_list);
+			if (cdata->fd > 0) {
+				usbmuxd_socket_close(cdata->fd);
+			}
+			free(cdata);
+			return NULL;
+		}
+
+		if (dev_list == NULL || dev_list[0].handle == 0) {
+			printf("No connected device found, terminating.\n");
+			free(dev_list);
+			if (cdata->fd > 0) {
+				usbmuxd_socket_close(cdata->fd);
+			}
+			free(cdata);
+			return NULL;
+		}
+
 		int i;
 		for (i = 0; i < count; i++) {
-			if (strncmp(dev_list[i].udid, device_udid, sizeof(dev_list[0].udid)) == 0) {
+			if (dev_list[i].conn_type == CONNECTION_TYPE_USB && (lookup_opts & DEVICE_LOOKUP_USBMUX)) {
+				dev = &(dev_list[i]);
+				break;
+			} else if (dev_list[i].conn_type == CONNECTION_TYPE_NETWORK && (lookup_opts & DEVICE_LOOKUP_NETWORK)) {
 				dev = &(dev_list[i]);
 				break;
 			}
 		}
-	} else {
-		dev = &(dev_list[0]);
 	}
 
 	if (dev == NULL || dev->handle == 0) {
 		printf("No connected/matching device found, disconnecting client.\n");
 		free(dev_list);
 		if (cdata->fd > 0) {
-			socket_close(cdata->fd);
+			usbmuxd_socket_close(cdata->fd);
 		}
 		free(cdata);
 		return NULL;
 	}
 
-	fprintf(stdout, "Requesting connecion to device handle == %d (serial: %s), port %d\n", dev->handle, dev->udid, device_port);
+	cdata->sfd = -1;
+	if (dev->conn_type == CONNECTION_TYPE_NETWORK) {
+		unsigned char saddr_[32];
+		memset(saddr_, '\0', sizeof(saddr_));
+		struct sockaddr* saddr = (struct sockaddr*)&saddr_[0];
+		if (((char*)dev->conn_data)[1] == 0x02) { // AF_INET
+			saddr->sa_family = AF_INET;
+			memcpy(&saddr->sa_data[0], (char*)dev->conn_data+2, 14);
+		}
+		else if (((char*)dev->conn_data)[1] == 0x1E) { //AF_INET6 (bsd)
+#ifdef AF_INET6
+			saddr->sa_family = AF_INET6;
+			memcpy(&saddr->sa_data[0], (char*)dev->conn_data+2, 26);
+#else
+			fprintf(stderr, "ERROR: Got an IPv6 address but this system doesn't support IPv6\n");
+			free(cdata);
+			return NULL;
+#endif
+		}
+		else {
+			fprintf(stderr, "Unsupported address family 0x%02x\n", ((char*)dev->conn_data)[1]);
+			free(cdata);
+			return NULL;
+		}
+		char addrtxt[48];
+		addrtxt[0] = '\0';
+		if (!usbmuxd_socket_addr_to_string(saddr, addrtxt, sizeof(addrtxt))) {
+			fprintf(stderr, "Failed to convert network address: %d (%s)\n", errno, strerror(errno));
+		}
+		fprintf(stdout, "Requesting connecion to NETWORK device %s (serial: %s), port %d\n", addrtxt, dev->udid, device_port);
+		cdata->sfd = usbmuxd_socket_connect_addr(saddr, device_port);
+	} else if (dev->conn_type == CONNECTION_TYPE_USB) {
+		fprintf(stdout, "Requesting connecion to USB device handle %d (serial: %s), port %d\n", dev->handle, dev->udid, device_port);
 
-	cdata->sfd = usbmuxd_connect(dev->handle, device_port);
+		cdata->sfd = usbmuxd_connect(dev->handle, device_port);
+	}
 	free(dev_list);
 	if (cdata->sfd < 0) {
-		fprintf(stderr, "Error connecting to device!\n");
+		fprintf(stderr, "Error connecting to device: %s\n", strerror(errno));
 	} else {
 		cdata->stop_ctos = 0;
 
@@ -242,56 +291,112 @@ static void *acceptor_thread(void *arg)
 	}
 
 	if (cdata->fd > 0) {
-		socket_close(cdata->fd);
+		usbmuxd_socket_close(cdata->fd);
 	}
 	if (cdata->sfd > 0) {
-		socket_close(cdata->sfd);
+		usbmuxd_socket_close(cdata->sfd);
 	}
 	free(cdata);
 
 	return NULL;
 }
 
+static void print_usage(int argc, char **argv, int is_error)
+{
+	char *name = NULL;
+	name = strrchr(argv[0], '/');
+	fprintf(is_error ? stderr : stdout, "Usage: %s [OPTIONS] LOCAL_PORT DEVICE_PORT\n", (name ? name + 1: argv[0]));
+	fprintf(is_error ? stderr : stdout,
+	  "Proxy that enables TCP service access to iOS devices.\n\n" \
+	  "  -u, --udid UDID    target specific device by UDID\n" \
+	  "  -n, --network      connect to network device\n" \
+	  "  -l, --local        connect to USB device (default)\n" \
+	  "  -h, --help         prints usage information\n" \
+	  "  -d, --debug        increase debug level\n" \
+	  "\n" \
+	  "Homepage: <" PACKAGE_URL ">\n"
+	  "Bug reports: <" PACKAGE_BUGREPORT ">\n"
+	  "\n"
+	);
+}
+
 int main(int argc, char **argv)
 {
 	int mysock = -1;
-
-	if (argc < 3) {
-		printf("usage: %s LOCAL_TCP_PORT DEVICE_TCP_PORT [UDID] [USBMUXD_HOST]\n", argv[0]);
-		printf("\n");
-		printf("LOCAL_TCP_PORT:\t\tThe TCP port to open on the host machine\n");
-		printf("DEVICE_TCP_PORT:\tThe TCP port to which to connect on the device\n");
-		printf("UDID:\t\t\tThe UDID of the device to which to connect\n");
-		printf("USBMUXD_HOST:\t\tThe The host on which usbmuxd is running. Can only be specified in combination with UDID. Forces TCP communication with usbmuxd.\n");
-		return 0;
+	const struct option longopts[] = {
+		{ "debug", no_argument, NULL, 'd' },
+		{ "help", no_argument, NULL, 'h' },
+		{ "udid", required_argument, NULL, 'u' },
+		{ "local", no_argument, NULL, 'l' },
+		{ "network", no_argument, NULL, 'n' },
+		{ NULL, 0, NULL, 0}
+	};
+	int c = 0;
+	while ((c = getopt_long(argc, argv, "dhu:lnm", longopts, NULL)) != -1) {
+		switch (c) {
+		case 'd':
+			libusbmuxd_set_debug_level(++debug_level);
+			break;
+		case 'u':
+			if (!*optarg) {
+				fprintf(stderr, "ERROR: UDID must not be empty!\n");
+				print_usage(argc, argv, 1);
+				return 2;
+			}
+			free(device_udid);
+			device_udid = strdup(optarg);
+			break;
+		case 'l':
+			lookup_opts |= DEVICE_LOOKUP_USBMUX;
+			break;
+		case 'n':
+			lookup_opts |= DEVICE_LOOKUP_NETWORK;
+			break;
+		case 'h':
+			print_usage(argc, argv, 0);
+			return 0;
+		default:
+			print_usage(argc, argv, 1);
+			return 2;
+		}
 	}
 
-	listen_port = atoi(argv[1]);
-	device_port = atoi(argv[2]);
-
-	if (argc > 3) {
-		device_udid = argv[3];
+	if (lookup_opts == 0) {
+		lookup_opts = DEVICE_LOOKUP_USBMUX;
 	}
 
-	if (argc > 4) {
-		usbmuxd_set_socket_type(SOCKET_TYPE_TCP);
-		usbmuxd_set_tcp_endpoint(argv[4], USBMUXD_SOCKET_PORT);
+	argc -= optind;
+	argv += optind;
+
+	if (argc < 2) {
+		print_usage(argc + optind, argv - optind, 1);
+		free(device_udid);
+		return 2;
 	}
+
+	listen_port = atoi(argv[0]);
+	device_port = atoi(argv[1]);
 
 	if (!listen_port) {
-		fprintf(stderr, "Invalid listen_port specified!\n");
+		fprintf(stderr, "Invalid listen port specified!\n");
+		free(device_udid);
 		return -EINVAL;
 	}
 
 	if (!device_port) {
-		fprintf(stderr, "Invalid device_port specified!\n");
+		fprintf(stderr, "Invalid device port specified!\n");
+		free(device_udid);
 		return -EINVAL;
 	}
 
+#ifndef WIN32
+	signal(SIGPIPE, SIG_IGN);
+#endif
 	// first create the listening socket endpoint waiting for connections.
-	mysock = socket_create(listen_port);
+	mysock = usbmuxd_socket_create(listen_port);
 	if (mysock < 0) {
 		fprintf(stderr, "Error creating socket: %s\n", strerror(errno));
+		free(device_udid);
 		return -errno;
 	} else {
 #ifdef WIN32
@@ -303,13 +408,14 @@ int main(int argc, char **argv)
 		int c_sock;
 		while (1) {
 			printf("waiting for connection\n");
-			c_sock = socket_accept(mysock, listen_port);
+			c_sock = usbmuxd_socket_accept(mysock, listen_port);
 			if (c_sock) {
 				printf("accepted connection, fd = %d\n", c_sock);
 				cdata = (struct client_data*)malloc(sizeof(struct client_data));
 				if (!cdata) {
-					socket_close(c_sock);
+					usbmuxd_socket_close(c_sock);
 					fprintf(stderr, "ERROR: Out of memory\n");
+					free(device_udid);
 					return -1;
 				}
 				cdata->fd = c_sock;
@@ -324,9 +430,11 @@ int main(int argc, char **argv)
 				break;
 			}
 		}
-		socket_close(c_sock);
-		socket_close(mysock);
+		usbmuxd_socket_close(c_sock);
+		usbmuxd_socket_close(mysock);
 	}
+
+	free(device_udid);
 
 	return 0;
 }
