@@ -1,6 +1,6 @@
 /*
  obs-ios-camera-source
- Copyright (C) 2018    Will Townsend <will@townsend.io>
+ Copyright (C) 2018 Will Townsend <will@townsend.io>
  
  This program is free software; you can redistribute it and/or modify
  it under the terms of the GNU General Public License as published by
@@ -16,58 +16,108 @@
  with this program. If not, see <https://www.gnu.org/licenses/>
  */
 
+#pragma once
+
 #include <chrono>
 
-#include "obs-ios-camera-source.h"
 #include "VideoDecoder.h"
 #include "ffmpeg-decode.h"
 #include "Queue.hpp"
 #include "Thread.hpp"
 
-class Decoder
-{
-    struct ffmpeg_decode decode;
-    
+class Decoder {
+	struct ffmpeg_decode decode;
+
 public:
-    inline Decoder() { memset(&decode, 0, sizeof(decode)); }
-    inline ~Decoder() { ffmpeg_decode_free(&decode); }
-    
-    inline operator ffmpeg_decode *() { return &decode; }
-    inline ffmpeg_decode *operator->() { return &decode; }
+	inline Decoder() { memset(&decode, 0, sizeof(decode)); }
+	inline ~Decoder() { ffmpeg_decode_free(&decode); }
+
+	inline operator ffmpeg_decode *() { return &decode; }
+	inline ffmpeg_decode *operator->() { return &decode; }
 };
 
-class FFMpegVideoDecoderCallback {
-public:
-    virtual ~FFMpegVideoDecoderCallback() {}
-    
-//    virtual void VideoToolboxDecodedFrame(CVPixelBufferRef aImage, CMVideoFormatDescriptionRef formatDescription) = 0;
-};
 
-class FFMpegVideoDecoder: public VideoDecoder, private Thread
-{
+class FFMpegVideoDecoder
+	: public VideoDecoder,
+	  private Thread,
+	  public std::enable_shared_from_this<FFMpegVideoDecoder> {
 public:
-    FFMpegVideoDecoder();
-    ~FFMpegVideoDecoder();
-    
-    void Init() override;
-    
-    void Input(std::vector<char> packet) override;
-    
-    void Flush() override;
-    void Drain() override;
-    void Shutdown() override;
-    
-    obs_source_t *source;
+	class Delegate {
+	public:
+		virtual void ffmpegDecoderDidDecodeFrame(void *data, int stride,
+							 int width,
+							 int height) = 0;
+		virtual ~Delegate(){};
+	};
+
+	FFMpegVideoDecoder();
+	~FFMpegVideoDecoder();
+
+	void Init() override;
+
+	void Input(std::vector<char> packet, int type, int tag) override;
+
+	void Flush() override;
+	void Drain() override;
+	void Shutdown() override;
+
+	void setDelegate(std::shared_ptr<Delegate> newDelegate)
+	{
+		delegate = newDelegate;
+	}
+
+	std::weak_ptr<Delegate> getDelegate() { return delegate; };
+
+	    obs_source_t *source;
+	obs_source_frame video_frame;
+
 
 private:
-    
-    void *run() override;
-    
-    void processPacketItem(PacketItem *packetItem);
-    
-    WorkQueue<PacketItem *> mQueue;
-    
-    obs_source_frame frame;
-    Decoder video_decoder;
-    
+	void *run() override;
+
+	void processPacketItem(PacketItem *packetItem);
+
+	WorkQueue<PacketItem *> mQueue;
+
+	Decoder video_decoder;
+	std::weak_ptr<Delegate> delegate;
+	std::mutex mMutex;
+
 };
+
+//
+//class FFMpegVideoDecoderCallback {
+//public:
+//    virtual ~FFMpegVideoDecoderCallback() {}
+//};
+//
+//class FFMpegVideoDecoder: public VideoDecoder, private Thread
+//{
+//public:
+//    FFMpegVideoDecoder();
+//    ~FFMpegVideoDecoder();
+//    
+//    void Init() override;
+//    
+//    void Input(std::vector<char> packet, int type, int tag) override;
+//    
+//    void Flush() override;
+//    void Drain() override;
+//    void Shutdown() override;
+//    
+//    obs_source_t *source;
+//
+//private:
+//    
+//    void *run() override;
+//    
+//    void processPacketItem(PacketItem *packetItem);
+//    
+//    WorkQueue<PacketItem *> mQueue;
+//    
+//    obs_source_frame video_frame;
+//    
+//    Decoder video_decoder;
+//
+//    std::mutex mMutex;
+//};
